requirements.txt
discord.py[voice]>=2.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
edge-tts>=6.1.0
